pass dtype for sim data

+++ larch/util/simdata_util.py
@@ -214,7 +214,7 @@ def sim_real(N, bulk_file, outfile, noise=5, seed = 123):
             ).sample()
         ).sample()
 
-    adata = ad.AnnData(X)
+    adata = ad.AnnData(X, dtype=X.dtype)
 
     adata.obs = pd.DataFrame({
         "cell_type": cell_type_names,
